[MOVE] - #29 NickNameCheckVC 에서 SignUpPopUP 으로 코드 이동
