+++ HACKER/HACKER/Sources/NetworkService/SignUp/SignUpAPI.swift
@@ -1,0 +1,71 @@
+//
+//  SignUpAPI.swift
+//  HACKER
+//
+//  Created by 김지수 on 2022/05/05.
+//
+
+import Foundation
+import Moya
+
+public class SignUpAPI {
+
+    static let shared = SignUpAPI()
+    var fightProvider = MoyaProvider<SignUpService>(plugins: [MoyaLoggerPlugin()])
+
+    public init() { }
+
+    func userGithubName(completion: @escaping (NetworkResult<Any>) -> Void) {
+      fightProvider.request(.userGithubName) { (result) in
+            switch result {
+            case .success(let response):
+                let statusCode = response.statusCode
+                let data = response.data
+
+                let networkResult = self.judgeStatus(by: statusCode, data)
+                completion(networkResult)
+                
+            case .failure(let err):
+                print(err)
+            }
+        }
+    }
+    
+    private func judgeGroupListFetchStatus(by statusCode: Int, _ data: Data) -> NetworkResult<Any> {
+
+        let decoder = JSONDecoder()
+        guard let decodedData = try? decoder.decode(GenericResponse<SeasonResponse>.self, from: data)
+        else {
+            return .pathErr
+        }
+
+        switch statusCode {
+        case 200:
+            return .success(decodedData.data ?? "None-Data")
+        case 400..<500:
+          return .requestErr(decodedData.message)
+        case 500:
+            return .serverErr
+        default:
+            return .networkFail
+        }
+    }
+    
+    private func judgeStatus(by statusCode: Int, _ data: Data) -> NetworkResult<Any> {
+        
+        let decoder = JSONDecoder()
+        guard let decodedData = try? decoder.decode(GenericResponse<String>.self, from: data)
+        else { return .pathErr }
+        
+        switch statusCode {
+        case 200:
+            return .success(decodedData.message)
+        case 400..<500:
+            return .requestErr(decodedData.message)
+        case 500:
+            return .serverErr
+        default:
+            return .networkFail
+        }
+    }
+}
